mini-project/pyproject.toml
[tool.poetry]
name = "dungeon-assistant"
version = "0.1.0"
description = ""
authors = ["Jonas Gerber <73141623+Traenqui@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
openai = "^1.52.2"
python-socketio = "^5.11.4"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
